+++ sdk/resourcemanager/agrifood/armagrifood/zz_generated_extensions_client.go
@@ -12,14 +12,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
 	"net/url"
 	"strconv"
 	"strings"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // ExtensionsClient contains the methods for the Extensions group.
@@ -31,8 +32,15 @@ type ExtensionsClient struct {
 }
 
 // NewExtensionsClient creates a new instance of ExtensionsClient with the specified values.
-func NewExtensionsClient(con *arm.Connection, subscriptionID string) *ExtensionsClient {
-	return &ExtensionsClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
+func NewExtensionsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *ExtensionsClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &ExtensionsClient{subscriptionID: subscriptionID, ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // Create - Install extension.
@@ -86,7 +94,7 @@ func (client *ExtensionsClient) createCreateRequest(ctx context.Context, extensi
 func (client *ExtensionsClient) createHandleResponse(resp *http.Response) (ExtensionsCreateResponse, error) {
 	result := ExtensionsCreateResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Extension); err != nil {
-		return ExtensionsCreateResponse{}, err
+		return ExtensionsCreateResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -215,7 +223,7 @@ func (client *ExtensionsClient) getCreateRequest(ctx context.Context, extensionI
 func (client *ExtensionsClient) getHandleResponse(resp *http.Response) (ExtensionsGetResponse, error) {
 	result := ExtensionsGetResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Extension); err != nil {
-		return ExtensionsGetResponse{}, err
+		return ExtensionsGetResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -293,7 +301,7 @@ func (client *ExtensionsClient) listByFarmBeatsCreateRequest(ctx context.Context
 func (client *ExtensionsClient) listByFarmBeatsHandleResponse(resp *http.Response) (ExtensionsListByFarmBeatsResponse, error) {
 	result := ExtensionsListByFarmBeatsResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.ExtensionListResponse); err != nil {
-		return ExtensionsListByFarmBeatsResponse{}, err
+		return ExtensionsListByFarmBeatsResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -362,7 +370,7 @@ func (client *ExtensionsClient) updateCreateRequest(ctx context.Context, extensi
 func (client *ExtensionsClient) updateHandleResponse(resp *http.Response) (ExtensionsUpdateResponse, error) {
 	result := ExtensionsUpdateResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.Extension); err != nil {
-		return ExtensionsUpdateResponse{}, err
+		return ExtensionsUpdateResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }